+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0
 
 - Added `getParentRef` to be able to get the parent of nodes.
+- Added `getComputedProp` to be able to generated computed functions that take an object and get a computed value related to that object without having to extend the object itself.
 
 ## 1.0.1